DjangoWebsite/requirements.txt
asgiref==3.6.0
attrs==22.2.0
colorama==0.4.6
confusable-homoglyphs==3.2.0
crispy-bootstrap5==0.7
Django==4.1.7
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-registration==3.3
django-social-share==2.3.0
django-tinymce==3.5.0
exceptiongroup==1.1.0
gunicorn==20.1.0
iniconfig==2.0.0
install==1.3.5
packaging==23.0
Pillow==9.4.0
pip==23.0.1
pytest==7.2.1
pytz==2022.7.1
pluggy==1.0.0
setuptools==67.3.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
wfastcgi==3.0.0
whitenoise==6.4.0
